Implementation of alphabeta + quiescence tree search + fix of unrealised bug in the code of en passant

+++ Board/Bishop.cpp
@@ -1,38 +1,40 @@
 #include "Bishop.h"
 
-void Bishop::add_moves(Board& board, short int player_move_multiplier, bool check_for_pin, BoardCell& board_cell)
+void Bishop::add_moves(Board &board, int player_move_multiplier, bool check_for_pin, BoardCell &board_cell)
 {
     Coord piece_c = this->piece_coord;
-    short int check_x{};
-    short int check_y{};
-    BoardCell* check_cell{ nullptr };
+    int check_x{};
+    int check_y{};
+    BoardCell *check_cell{nullptr};
 
     // checks to the top right of the bishop for possible moves
-    for (short int i = 1; i < board_size + 1; i++)
+    for (int i = 1; i < board_size + 1; i++)
     {
         check_x = piece_c.x + i;
         check_y = piece_c.y + i;
-        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size) {
+        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size)
+        {
             check_cell = &board.board[check_x][check_y];
-            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner) {
-                Coord move_to{ check_x, check_y };
+            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner)
+            {
+                Coord move_to{check_x, check_y};
                 board.add_move(move_to, piece_c, player_move_multiplier, check_for_pin);
-                if (check_cell->piece != nullptr) { break; }
-            }
-            else { break; }
-        }
-        else { break; }
+                if (check_cell->piece != nullptr){ break;}
+            } else { break; }
+        } else { break; }
     }
 
     // checks to the top left of the bishop for possible moves
-    for (short int i = 1; i < board_size + 1; i++)
+    for (int i = 1; i < board_size + 1; i++)
     {
         check_x = piece_c.x - i;
         check_y = piece_c.y + i;
-        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size) {
+        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size)
+        {
             check_cell = &board.board[check_x][check_y];
-            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner) {
-                Coord move_to{ check_x, check_y };
+            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner)
+            {
+                Coord move_to{check_x, check_y};
                 board.add_move(move_to, piece_c, player_move_multiplier, check_for_pin);
                 if (check_cell->piece != nullptr) { break; }
             }
@@ -42,14 +44,16 @@ void Bishop::add_moves(Board& board, short int player_move_multiplier, bool chec
     }
 
     // checks to the bottom right of the bishop for possible moves
-    for (short int i = 1; i < board_size+1; i++)
+    for (int i = 1; i < board_size + 1; i++)
     {
         check_x = piece_c.x + i;
         check_y = piece_c.y - i;
-        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size) {
+        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size)
+        {
             check_cell = &board.board[check_x][check_y];
-            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner) {
-                Coord move_to{ check_x, check_y };
+            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner)
+            {
+                Coord move_to{check_x, check_y};
                 board.add_move(move_to, piece_c, player_move_multiplier, check_for_pin);
                 if (check_cell->piece != nullptr) { break; }
             }
@@ -59,14 +63,16 @@ void Bishop::add_moves(Board& board, short int player_move_multiplier, bool chec
     }
 
     // checks to the bottom left of the bishop for possible moves
-    for (short int i = 1; i < board_size + 1; i++)
+    for (int i = 1; i < board_size + 1; i++)
     {
         check_x = piece_c.x - i;
         check_y = piece_c.y - i;
-        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size) {
+        if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size)
+        {
             check_cell = &board.board[check_x][check_y];
-            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner) {
-                Coord move_to{ check_x, check_y };
+            if (check_cell->piece == nullptr || check_cell->piece != nullptr && check_cell->piece->owner != board_cell.piece->owner)
+            {
+                Coord move_to{check_x, check_y};
                 board.add_move(move_to, piece_c, player_move_multiplier, check_for_pin);
                 if (check_cell->piece != nullptr) { break; }
             }
