+++ Board/Knight.cpp
@@ -1,14 +1,14 @@
 #include "Knight.h"
 
-short int knight_x[8] = { 2, 1, -1, -2, -2, -1, 1, 2 };
-short int knight_y[8] = { 1, 2, 2, 1, -1, -2, -2, -1 };
+int knight_x[8] = { 2, 1, -1, -2, -2, -1, 1, 2 };
+int knight_y[8] = { 1, 2, 2, 1, -1, -2, -2, -1 };
 
-void Knight::add_moves(Board& board, short int player_move_multiplier, bool check_for_pin, BoardCell& board_cell) {
+void Knight::add_moves(Board& board, int player_move_multiplier, bool check_for_pin, BoardCell& board_cell) {
     Coord piece_c = this->piece_coord;
-    short int check_x{};
-    short int check_y{};
+    int check_x{};
+    int check_y{};
 
-    for (short int i = 0; i < board_size; i++) {
+    for (int i = 0; i < board_size; i++) {
         check_x = piece_c.x + knight_x[i];
         check_y = piece_c.y + knight_y[i];
         if (check_x > -1 && check_x < board_size && check_y > -1 && check_y < board_size) {
