+++ Board/Knight.h
@@ -4,8 +4,8 @@
 
 class Knight : public Piece {
 public:
-	Knight(const short int piece_type_param, const short int owner_param, Coord& coord_param) : Piece(piece_type_param, owner_param, coord_param) {}
+	Knight(const int piece_type_param, const int owner_param, Coord& coord_param) : Piece(piece_type_param, owner_param, coord_param) {}
 
-	void add_moves(Board& board, short int player_move_multiplier, bool check_for_pin, BoardCell& board_cell);
+	void add_moves(Board& board, int player_move_multiplier, bool check_for_pin, BoardCell& board_cell);
 };
