+++ Board/Piece.h
@@ -1,12 +1,12 @@
 #pragma once
 #include "Move.h"
 
-const short int p = 100;
-const short int b = 305;
-const short int n = 300;
-const short int r = 500;
-const short int q = 900;
-const short int k = 2000;
+const int p = 100;
+const int b = 305;
+const int n = 300;
+const int r = 500;
+const int q = 900;
+const int k = 2000;
 
 class Board;
 class BoardCell;
@@ -14,12 +14,12 @@ class BoardCell;
 class Piece
 {
 public:
-	Piece(const short int piece_type_param, const short int owner_param, Coord& coord_param);
+	Piece(const int piece_type_param, const int owner_param, Coord& coord_param);
 	~Piece();
 
 	//board_cell.piece->add_moves(*this, player_move_multiplier, check_for_pin);
-	virtual void add_moves(Board& board, short int player_move_multiplier, bool check_for_pin, BoardCell& board_cell){}
+	virtual void add_moves(Board& board, int player_move_multiplier, bool check_for_pin, BoardCell& board_cell){}
 	Coord piece_coord{};
-	short int owner{};
-	short int piece_type{};
+	int owner{};
+	int piece_type{};
 };
